+++ src/core-interfaces.ts
@@ -1227,8 +1227,22 @@ export interface IChartPropsChartBar {
 	bar3DShape?: string
 	barDir?: string
 	barGapDepthPct?: number
+	/**
+	 * MS-PPT > Format chart > Format Data Point > Series Options >  "Gap Width"
+	 * - width (percent)
+	 * - range: `0`-`500`
+	 * @default 150
+	 */
 	barGapWidthPct?: number
 	barGrouping?: string
+	/**
+	 * MS-PPT > Format chart > Format Data Point > Series Options >  "Series Overlap"
+	 * - overlap (percent)
+	 * - range: `-100`-`100`
+	 * @since v3.9.0
+	 * @default 0
+	 */
+	barOverlapPct?: number
 }
 export interface IChartPropsChartDoughnut {
 	dataNoEffects?: boolean